+++ src/app/modules/data-selection/components/editor/display/data-selection-boxes/data-selection-boxes.component.ts
@@ -10,15 +10,18 @@ import { FilterChipData } from 'src/app/shared/models/FilterChips/FilterChipData
 import { FilterChipsComponent } from '../../../../../../shared/components/filter-chips/filter-chips.component'
 import { InfoTooltipDirective } from '../../../../../../shared/directives/info-tooltip.directive'
 import { LinkedBadgeComponent } from '../../../../../../shared/components/linked-badge/linked-badge.component'
-import { map } from 'rxjs'
+import { map, tap } from 'rxjs'
 import { MenuComponent } from '../../../../../../shared/components/menu/menu.component'
 import { MenuItemInterface } from '../../../../../../shared/models/Menu/MenuItemInterface'
 import { MenuServiceDataSelection } from '../../../../../../shared/service/Menu/DataSelection/MenuServiceDataSelection.service'
+import { MissingFilterComponent } from '../../../../../../shared/components/missing-filter/missing-filter.component'
+import { NavigationHelperService } from 'src/app/service/NavigationHelper.service'
 import { ProfileReference } from 'src/app/model/DataSelection/Profile/Reference/ProfileReference'
 import { ProfileReferenceTileComponent } from '../../../../../../shared/components/profile-reference-tile/profile-reference-tile.component'
 import { RemoveReferenceService } from 'src/app/service/RemoveReference.service'
 import { toSignal } from '@angular/core/rxjs-interop'
 import { TranslateModule } from '@ngx-translate/core'
+import { MatTooltip } from '@angular/material/tooltip'
 
 @Component({
   selector: 'num-data-selection-boxes',
@@ -35,6 +38,8 @@ import { TranslateModule } from '@ngx-translate/core'
     ProfileReferenceTileComponent,
     TranslateModule,
     DisplayTranslationPipe,
+    MissingFilterComponent,
+    MatTooltip,
   ],
 })
 export class DataSelectionBoxesComponent {
@@ -44,6 +49,7 @@ export class DataSelectionBoxesComponent {
   private menuService = inject(MenuServiceDataSelection)
   private appSettingsProvider = inject(AppSettingsProviderService)
   private dataSelectionProviderService = inject(DataSelectionProviderService)
+  private navigationHelperService = inject(NavigationHelperService)
 
   readonly profile = input<DataSelectionProfile>()
   readonly isEditable = input<boolean>()
@@ -65,11 +71,11 @@ export class DataSelectionBoxesComponent {
   })
 
   readonly unlinkedRequiredOrRecommendedReferences = computed(
-    () => this.profile()?.getProfileFields().getUnlinkedRequiredOrRecommendedReferences() ?? []
+    () => this.profile().getProfileFields().getUnlinkedRequiredOrRecommendedReferences() ?? []
   )
 
   readonly selectedReferenceFields = computed(
-    () => this.profile()?.getProfileFields().getSelectedReferenceFields() ?? []
+    () => this.profile().getProfileFields().getSelectedReferenceFields() ?? []
   )
 
   readonly menuItems = computed<MenuItemInterface[]>(() => {
@@ -80,23 +86,29 @@ export class DataSelectionBoxesComponent {
 
   private readonly activeDataSelection$ = this.dataSelectionProviderService.getActiveDataSelection()
 
-  readonly isReferenced = toSignal(
-    this.activeDataSelection$.pipe(
-      map((dataSelection) =>
-        dataSelection.getProfiles().some((profile) =>
-          profile
-            .getProfileFields()
-            .getSelectedReferenceFields()
-            .some((referenceField) =>
-              referenceField
-                .getLinkedProfileIds()
-                .some((linkedProfileId) => this.profile()?.getId() === linkedProfileId)
-            )
-        )
-      )
-    ),
-    { initialValue: false }
-  )
+  readonly activeDataSelection = toSignal(this.activeDataSelection$, {
+    initialValue: undefined,
+  })
+
+  readonly isReferenced = computed(() => {
+    const dataSelection = this.activeDataSelection()
+    const profile = this.profile()
+
+    if (!dataSelection || !profile) {
+      return false
+    }
+
+    return dataSelection.getProfiles().some((selectionProfile) =>
+      selectionProfile
+        .getProfileFields()
+        .getSelectedReferenceFields()
+        .some((selectedReferenceField) => {
+          return selectedReferenceField
+            .getLinkedProfileIds()
+            .some((linkedProfileId) => profile.getId() === linkedProfileId)
+        })
+    )
+  })
 
   constructor() {
     effect(() => {
@@ -111,5 +123,9 @@ export class DataSelectionBoxesComponent {
 
   public deleteProfile(id: string): void {
     this.removeReferenceService.delete(id)
+  }
+
+  public navigate(): void {
+    this.navigationHelperService.navigateToEditProfile(this.profile().getId())
   }
 }