+++ src/app/modules/feasibility-query/components/editor/display/display-criteria/display-criteria.component.ts
@@ -30,7 +30,6 @@ import { TranslateModule } from '@ngx-translate/core'
     FontAwesomeModule,
     AsyncPipe,
     TranslateModule,
-    NgClass,
   ],
 })
 export class DisplayCriteriaComponent implements OnInit, OnDestroy {