+++ src/app/modules/feasibility-query/components/search/search.component.spec.ts
@@ -1,22 +1,22 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing'
 
-import { SearchComponent } from './search.component';
+import { FeasibilityQuerySearchComponent } from './search.component'
 
-describe('SearchComponent', () => {
-  let component: SearchComponent;
-  let fixture: ComponentFixture<SearchComponent>;
+describe('FeasibilityQuerySearchComponent', () => {
+  let component: FeasibilityQuerySearchComponent
+  let fixture: ComponentFixture<FeasibilityQuerySearchComponent>
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [SearchComponent],
-    }).compileComponents();
+      imports: [FeasibilityQuerySearchComponent],
+    }).compileComponents()
 
-    fixture = TestBed.createComponent(SearchComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+    fixture = TestBed.createComponent(FeasibilityQuerySearchComponent)
+    component = fixture.componentInstance
+    fixture.detectChanges()
+  })
 
   it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+    expect(component).toBeTruthy()
+  })
+})