+++ src/app/modules/query-editor/components/editor-content/filter-tabs/filter-tabs.component.ts
@@ -1,7 +1,8 @@
-import { Component, OnInit, TemplateRef, input } from '@angular/core'
+import { Component, input, OnInit, TemplateRef } from '@angular/core'
+import { Display } from 'src/app/model/DataSelection/Profile/Display'
+import { DisplayTranslationPipe } from '../../../../../shared/pipes/DisplayTranslationPipe'
 import { NgTemplateOutlet } from '@angular/common'
 import { TranslateModule } from '@ngx-translate/core'
-import { DisplayTranslationPipe } from '../../../../../shared/pipes/DisplayTranslationPipe'
 
 @Component({
   selector: 'num-filter-tabs',
@@ -11,10 +12,22 @@ import { DisplayTranslationPipe } from '../../../../../shared/pipes/DisplayTrans
   imports: [NgTemplateOutlet, TranslateModule, DisplayTranslationPipe],
 })
 export class FilterTabsComponent implements OnInit {
-  readonly content = input<{ template: TemplateRef<any>; name: string; context?: any }[]>()
+  readonly content = input.required<
+    {
+      template: TemplateRef<any>
+      name: string
+      context?: any
+      display?: Display
+      active?: boolean
+    }[]
+  >()
+
   selectedIndex = 0
 
   constructor() {}
 
-  ngOnInit(): void {}
+  ngOnInit(): void {
+    const activeTabIndex = this.content().findIndex((tab) => tab.active)
+    this.selectedIndex = activeTabIndex !== -1 ? activeTabIndex : 0
+  }
 }