+++ src/app/service/NavigationHelper.service.ts
@@ -32,8 +32,8 @@ export class NavigationHelperService {
    * @param id - The query profile identifier
    * @returns
    */
-  public navigateToEditProfile(id: string): void {
-    this.router.navigate([`${UrlPaths.queryEditor.feature}`, id])
+  public navigateToEditProfile(id: string, state?: { activeTab: string }): void {
+    this.router.navigate([`${UrlPaths.queryEditor.feature}`, id], { state })
   }
 
   public navigateToEditCriterion(id: string): void {
@@ -112,5 +112,9 @@ export class NavigationHelperService {
     this.router.navigate([BasePaths.savedQueries], {
       state: { preventReset: true },
     })
+  }
+
+  public getCurrentNavigation(): string {
+    return this.router.getCurrentNavigation()?.extras.state?.['activeTab']
   }
 }